move AssertionGeneratorUtils
