+++ dspot/src/main/java/eu/stamp_project/dspot/assertiongenerator/assertiongenerator_components/AssertionRemover.java
@@ -1,6 +1,5 @@
 package eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components;
 
-import eu.stamp_project.dspot.assertiongenerator.utils.AssertionGeneratorUtils;
 import eu.stamp_project.test_framework.TestFramework;
 import eu.stamp_project.utils.CloneHelper;
 import spoon.reflect.code.*;
