+++ dspot/src/test/java/eu/stamp_project/dspot/assertiongenerator/utils/AssertionGeneratorUtilsTest.java
@@ -2,7 +2,7 @@ package eu.stamp_project.dspot.assertiongenerator.utils;
 
 import eu.stamp_project.AbstractTest;
 import eu.stamp_project.Utils;
-import eu.stamp_project.dspot.assertiongenerator.utils.AssertionGeneratorUtils;
+import eu.stamp_project.dspot.assertiongenerator.assertiongenerator_components.AssertionGeneratorUtils;
 import eu.stamp_project.test_framework.TestFramework;
 import eu.stamp_project.utils.program.InputConfiguration;
 import org.junit.Test;
